pyproject.toml
[tool.poetry]
name = "Online Offline Researcher"
version = "0.0.1"
description = "Research a given topic using ollama local models."
authors = ["Anthony Sun <auslei@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
wikipedia = "^1.4.0"
beautifulsoup4 = "^4.12.3"
selenium = "^4.23.1"
transformers = "^4.44.2"
ollama = "^0.3.2"
tqdm = "^4.66.5"
langchain = "^0.2.0"
chromadb = "^0.5.5"
colorama = "^0.4.6"
tiktoken = "^0.7.0"
langchain-community = "^0.2.16"
pandas = "^2.2.2"
streamlit = "^1.38.0"
requests-html = "^0.10.0"
lxml-html-clean = "^0.2.2"
html2text = "^2024.2.26"
openai = "^1.51.0"
paperclip = "^2.7.2"
google-api-python-client = "^2.147.0"
newsapi-python = "^0.2.7"
sentence-transformers = "^3.2.0"
feedparser = "^6.0.11"
duckduckgo-search = "^6.3.2"
poetry-dotenv = "^0.4.0"
ipykernel = "^6.29.5"
json5 = "^0.9.25"
deep-translator = "^1.11.4"
yfinance = "^0.2.48"
pymupdf = "^1.24.13"
pymupdf4llm = "^0.0.17"
markdown-it-py = "^3.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.pyright] # check python code for type errors
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"